Fix for varriable size children as well as wrap_content parent
The onMeassure() step now finds the largest child when setting the
wrapping dimension.

measureChild() now properly handles the case where all parent views
are wrap_content. RecyclerView's own getChildMeasureSpec() has some
limitations in this regaurd which we must work around.

The result is allowing RecyclerView to size it's self entirely based
on it's children as well as deal with heterogenous children.